Remove unnecessary components from BasicInfoStep
