--- packages/admin-app/src/components/StepHeader/index.ts
@@ -1,5 +1,0 @@
-import StepHeader from './StepHeader';
-
-export * from './StepHeader';
-
-export default StepHeader;
