--- packages/admin-app/src/components/StepIntroduction/index.ts
@@ -1,5 +1,0 @@
-import StepIntroduction from './StepIntroduction';
-
-export * from './StepIntroduction';
-
-export default StepIntroduction;
